Add tests on parsing label file into shared format
