+++ tests/resource/gt_files/mir1k_gt_file.pv
@@ -1,0 +1,8 @@
+0 
+57.9108 
+57.4161 
+57.174 
+59.7627 
+60.0442 
+60.3304 
+0
