+++ tests/resource/gt_files/tonas_gt_file.notes.Corrected
@@ -1,0 +1,5 @@
+46.000000
+0.233333, 0.133333, 53.459999, 0.000154
+0.400000, 0.183333, 58.459999, 0.000524
+0.600000, 0.150000, 58.459999, 0.001655
+0.766667, 0.183333, 58.459999, 0.002391
